replace search with docs
